requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pillow==10.4.0
PyJWT==2.8.0
python-telegram-bot==21.4
sniffio==1.3.1
sqlparse==0.5.1
